qlib_research/requirements.txt
# qlib_research dependencies

# Core
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
toml>=0.10.2

# Scientific
scipy>=1.7.0

# Data processing
h5py>=3.0.0

# qlib (bundled in qlib/ folder, no pip install needed)
# The qlib folder contains the library and is already included in this project